@@ -24,8 +24,8 @@
   "dependencies": {
     "graphiql": "^2.2.0",
     "graphql": "^16.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "parcel": "^2.5.0",
