@@ -15,7 +15,7 @@
     "graphiql": "^2.2.0",
     "graphql": "^16.4.0",
     "graphql-ws": "^5.5.5",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "regenerator-runtime": "^0.13.9"
   },
   "devDependencies": {
@@ -27,7 +27,7 @@
     "cross-env": "^7.0.2",
     "css-loader": "^6.7.3",
     "html-webpack-plugin": "^4.2.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "style-loader": "^3.3.1",
     "webpack": "5.75.0",
     "webpack-cli": "^5.0.1",
