@@ -6,8 +6,8 @@
     "@graphiql/toolkit": "^0.8.0",
     "graphql-language-service": "^5.1.0",
     "monaco-graphql": "^1.1.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "graphql": "^16.1.0",
     "jsonc-parser": "^3.0.0",
     "monaco-editor": "^0.31.0"
